// frontend/src/types/user.ts
export type LoginDTO = {
  username: string;
  password: string;
};

export type USERSPEC = {
  username: string;
  displayName: string;
  colorTheme: string;
  textColorOnTheme: string;
  textBackground: string;
  buttonScheme: string;
  hoverTheme: string;
  variantsName: string;
  textColor: string;
};